requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core

langchain-community
faiss-cpu
pypdf
sentence-transformers
selenium 
beautifulsoup4
PyPDF2
tavily-python
